updated window tests and renamed event interface
