+++ feature/event_test.go
@@ -7,6 +7,8 @@ import (
 	"github.com/goccy/go-json"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"gonum.org/v1/gonum/dsp/window"
+	"gonum.org/v1/gonum/floats"
 )
 
 func TestEventString(t *testing.T) {
@@ -387,7 +389,7 @@ func TestEventGenerate(t *testing.T) {
 		event       *Event
 		times       []time.Time
 		window      string
-		evSetting   EventSetting
+		evGenerator EventGenerator
 		wantLen     int
 		wantErr     bool
 		description string
@@ -397,7 +399,7 @@ func TestEventGenerate(t *testing.T) {
 			event:       NewEvent("test"),
 			times:       []time.Time{baseTime, baseTime.Add(24 * time.Hour), baseTime.Add(48 * time.Hour)},
 			window:      WindowRectangular,
-			evSetting:   &EventStandard{Start: baseTime.Add(12 * time.Hour), End: baseTime.Add(36 * time.Hour)},
+			evGenerator: &EventStandard{Start: baseTime.Add(12 * time.Hour), End: baseTime.Add(36 * time.Hour)},
 			wantLen:     3,
 			wantErr:     false,
 			description: "Event using standard setting",
@@ -407,7 +409,7 @@ func TestEventGenerate(t *testing.T) {
 			event:       NewEvent("weekend_test"),
 			times:       []time.Time{baseTime, baseTime.Add(24 * time.Hour)},
 			window:      WindowRectangular,
-			evSetting:   &EventWeekend{DurBefore: 0, DurAfter: 0},
+			evGenerator: &EventWeekend{DurBefore: 0, DurAfter: 0},
 			wantLen:     2,
 			wantErr:     false,
 			description: "Event using weekend setting",
@@ -416,11 +418,11 @@ func TestEventGenerate(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			if weekend, ok := tc.evSetting.(*EventWeekend); ok {
+			if weekend, ok := tc.evGenerator.(*EventWeekend); ok {
 				weekend.Validate()
 			}
 
-			got, err := tc.event.Generate(tc.times, tc.window, tc.evSetting)
+			got, err := tc.event.Generate(tc.times, tc.window, tc.evGenerator)
 
 			if tc.wantErr {
 				assert.Error(t, err)
@@ -429,6 +431,42 @@ func TestEventGenerate(t *testing.T) {
 
 			assert.NoError(t, err)
 			assert.Len(t, got, tc.wantLen)
+		})
+	}
+}
+
+func TestWindowFunc(t *testing.T) {
+	testData := map[string]struct {
+		name     string
+		expected func([]float64) []float64
+	}{
+		"bartlett hann":    {WindowBartlettHann, window.BartlettHann},
+		"blackman":         {WindowBlackman, window.Blackman},
+		"blackman harris":  {WindowBlackmanHarris, window.BlackmanHarris},
+		"blackman nuttall": {WindowBlackmanNuttall, window.BlackmanNuttall},
+		"flat top":         {WindowFlatTop, window.FlatTop},
+		"hamming":          {WindowHamming, window.Hamming},
+		"hann":             {WindowHann, window.Hann},
+		"lanczos":          {WindowLanczos, window.Lanczos},
+		"nuttall":          {WindowNuttall, window.Nuttall},
+		"rectangular":      {WindowRectangular, window.Rectangular},
+		"sine":             {WindowSine, window.Sine},
+		"triangular":       {WindowTriangular, window.Triangular},
+		"unknown":          {"unknown", window.Rectangular},
+	}
+
+	numPnts := 10
+	for name, td := range testData {
+		t.Run(name, func(t *testing.T) {
+			res := WindowFunc(td.name)
+
+			seqRes := make([]float64, numPnts)
+			floats.AddConst(1.0, seqRes)
+
+			seqExp := make([]float64, numPnts)
+			floats.AddConst(1.0, seqExp)
+
+			assert.Equal(t, td.expected(seqExp), res(seqRes))
 		})
 	}
 }
